Addressed some of the review comments.
