+++ tests/integration/main_test.go
@@ -34,22 +34,22 @@ func TestMain(m *testing.M) {
 func FeatureContext(s *godog.Suite) {
 
 	context := new(support.ScenarioCtx)
-
+	// Resetting context before each scenario
+	s.BeforeScenario(func(interface{}) {
+		context.Reset()
+	})
 	s.Step(`^the datafile is "([^"]*)"$`, context.TheDatafileIs)
 	s.Step(`^(\d+) "([^"]*)" listener is added$`, context.ListenerIsAdded)
 	s.Step(`^([^\\\"]*) is called with arguments$`, context.IsCalledWithArguments)
 	s.Step(`^the result should be (?:string )?"([^"]*)"$`, context.TheResultShouldBeString)
 	s.Step(`^the result should be (?:integer )?(\d+)$`, context.TheResultShouldBeInteger)
 	s.Step(`^the result should be (?:double )?(\d+)\.(\d+)$`, context.TheResultShouldBeFloat)
-	s.Step(`^the result should be boolean "([^"]*)"$`, context.TheResultShouldBeBoolean)
-	s.Step(`^the result should be \'false\'$`, context.TheResultShouldBeFalse)
+	s.Step(`^the result should be boolean "([^"]*)"$`, context.TheResultShouldBeTypedBoolean)
+	s.Step(`^the result should be \'([^"]*)\'$`, context.TheResultShouldBeBoolean)
 	s.Step(`^the result should match list "([^"]*)"$`, context.TheResultShouldMatchList)
 	s.Step(`^in the response, "([^"]*)" should be "([^"]*)"$`, context.InTheResponseKeyShouldBeObject)
 	s.Step(`^in the response, "([^"]*)" should match$`, context.InTheResponseShouldMatch)
 	s.Step(`^in the response, "([^"]*)" should have each one of these$`, context.InTheResponseShouldHaveEachOneOfThese)
 	s.Step(`^there are no dispatched events$`, context.ThereAreNoDispatchedEvents)
 	s.Step(`^dispatched events payloads include$`, context.DispatchedEventsPayloadsInclude)
-	s.BeforeScenario(func(interface{}) {
-		context.Reset()
-	})
 }
