+++ tests/integration/models/get_enabled_features_params.go
@@ -1,6 +1,6 @@
 package models
 
-// GetEnabledFeaturesParams represents params required for isFeatureEnabled API
+// GetEnabledFeaturesParams represents params required for GetEnabledFeatures API
 type GetEnabledFeaturesParams struct {
 	UserID     string                 `yaml:"user_id"`
 	Attributes map[string]interface{} `yaml:"attributes"`
