+++ tests/integration/models/get_feature_variable_params.go
@@ -1,6 +1,6 @@
 package models
 
-// GetFeatureVariableParams represents params required for isFeatureEnabled API
+// GetFeatureVariableParams represents params required for GetFeatureVariable API's
 type GetFeatureVariableParams struct {
 	FeatureKey  string                 `yaml:"feature_flag_key"`
 	VariableKey string                 `yaml:"variable_key"`
